+++ cli.py
@@ -1,4 +1,4 @@
-from modules.main import main
+from scrap_udemy_course.main import main
 
 if __name__ == "__main__":
     main()