+++ scrap_udemy_course/utils.py
@@ -30,7 +30,7 @@ from pathlib import Path
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 
-from modules import config
+from . import config
 
 
 # ============================================================================